+++ service-datapack/src/main/java/com/dataint/service/datapack/model/form/ArticleUpdateForm.java
@@ -13,11 +13,11 @@ public class ArticleUpdateForm {
 
     @ApiModelProperty(value = "舆情表-主键Id")
     @NotNull
-    private Integer articleId;
+    private Long articleId;
 
     @ApiModelProperty(value = "舆情等级表-主键Id")
     @NotNull
-    private Integer levelId;
+    private Long levelId;
 
     @ApiModelProperty(value = "舆情摘要")
     private String summary;
