+++ service-datapack/src/main/java/com/dataint/service/datapack/model/form/SourceUpdateForm.java
@@ -1,7 +1,7 @@
 package com.dataint.service.datapack.model.form;
 
 import com.dataint.cloud.common.model.form.BaseForm;
-import com.dataint.service.datapack.dao.entity.Source;
+import com.dataint.service.datapack.db.entity.Source;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
@@ -14,11 +14,11 @@ import javax.validation.constraints.NotNull;
 public class SourceUpdateForm extends BaseForm<Source> {
     @ApiModelProperty(value = "链接id")
     @NotNull
-    private Integer sourceId;
+    private Long sourceId;
 
     @NotNull(message = "国家Id不能为空")
     @ApiModelProperty(value = "国家Id")
-    private Integer countryId;
+    private Long countryId;
 
     @ApiModelProperty(value = "链接描述")
     private String description;
