+++ service-datapack/src/main/java/com/dataint/service/datapack/service/impl/SourceServiceImpl.java
@@ -1,13 +1,13 @@
 package com.dataint.service.datapack.service.impl;
 
 import com.dataint.cloud.common.exception.DataNotExistException;
-import com.dataint.service.datapack.dao.ICountryDao;
-import com.dataint.service.datapack.dao.ISourceDao;
-import com.dataint.service.datapack.dao.entity.Country;
-import com.dataint.service.datapack.dao.entity.Source;
-import com.dataint.service.datapack.model.SourceVO;
+import com.dataint.service.datapack.db.dao.ICountryDao;
+import com.dataint.service.datapack.db.dao.ISourceDao;
+import com.dataint.service.datapack.db.entity.Country;
+import com.dataint.service.datapack.db.entity.Source;
 import com.dataint.service.datapack.model.form.SourceForm;
 import com.dataint.service.datapack.model.form.SourceUpdateForm;
+import com.dataint.service.datapack.model.vo.SourceVO;
 import com.dataint.service.datapack.service.ISourceService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -58,7 +58,7 @@ public class SourceServiceImpl implements ISourceService {
     }
 
     @Override
-    public boolean delSource(Integer sourceId) {
+    public boolean delSource(Long sourceId) {
         // check if Source exist
         Optional<Source> sourceOpt = sourceDao.findById(sourceId);
         if (!sourceOpt.isPresent()) {
